Fixed bugs with duration and stopping of the system and implemented help and state commands

+++ controller/gpio_utils.h
@@ -16,7 +16,7 @@ int initGpioPinControl();
 bool checkDeviceState();
 
 //dispenses water for a certain amount of time
-void runIrrigation(int duration);
+void runIrrigation(float duration);
 
 //turns off the water pump
 void shutdownWaterPump();
